requirements.txt
asgiref==3.4.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
defusedxml==0.7.1
Django==3.2.19
django-allauth==0.53.1
django-cors-headers==3.10.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
Pillow==8.4.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2023.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1
urllib3==1.26.15


cryptography==40.0.2
